requirements.txt
requests
urllib3
tabulate
pillow
